Geopub_database: Add support for ρdf inference.
This adds inference of types by using RDF Schema ontologies.

For example:

ex:a a as:Note

will infer

ex:a a as:Object

because we have

as:Note rdfs:subClassOf as:Object

Concretely this allows queries for all activities while using
information from the ontology on what an activity is (and more).

However, this is not usable for unknown performance reasons.  We need to
optimize IndexedDB and the Datalog evaluation.

See also [Muñoz, S., Pérez, J., & Gutierrez, C. (2009). Simple and
Efficient Minimal RDFS. Web Semantics: Science, Services and Agents on
the World Wide Web, 7(3), 220–234. doi:10.1016/j.websem.2009.07.003].