Geopub_database.Store.Triples: Reduce number of indices by using IndexedDB range queries
This replaces the previous indices with three indices: spo, pos and
osp. With range queries these are sufficient to query RDF data in any
direction.

For example if we query for a fixed s we can use the "sop" index with
lower range [s] and upper range [s, ""]. The "" will always be sorted
higher than any integer and so we will get all entries for which the s
matches.